Fixed Bio crash on android 7/8

+++ lib/util/Utilities.dart
@@ -11,7 +11,6 @@ import 'package:notes/util/AppRoutes.dart';
 import 'package:notes/util/LockManager.dart';
 import 'package:notes/util/Navigations.dart';
 import 'package:notes/widget/AlertDialog.dart';
-import 'package:notes/widget/SimpleDialog.dart';
 import 'package:pedantic/pedantic.dart';
 import 'package:permission_handler/permission_handler.dart';
 import 'package:provider/provider.dart';
@@ -48,6 +47,7 @@ class Utilities {
   static late SharedPreferences prefs;
   static const passLength = 4;
   static const aboutMePic = 'me.png';
+  static const appName = 'KeepNotes';
 
   static late FlutterSecureStorage storage;
 
@@ -56,8 +56,6 @@ class Utilities {
     if (deleteAllNotes) {
       await Provider.of<NotesHelper>(context, listen: false)
           .deleteAllHiddenNotesHelper();
-      Provider.of<LockChecker>(context, listen: false).passwordSet = false;
-      Provider.of<LockChecker>(context, listen: false).updateDetails();
     } else {
       Utilities.showSnackbar(context, Language.of(context).done);
       unawaited(Provider.of<NotesHelper>(context, listen: false)
@@ -74,7 +72,8 @@ class Utilities {
       context,
       Language.of(context).passwordReset,
     );
-    await Provider.of<LockChecker>(context, listen: false).resetConfig();
+    await Provider.of<LockChecker>(context, listen: false)
+        .resetConfig(shouldResetBio: true);
     await navigate('', context, AppRoutes.homeScreen);
   }
 
@@ -118,31 +117,37 @@ class Utilities {
   static SnackBarAction resetAction(BuildContext context) => SnackBarAction(
         label: Language.of(context).reset,
         onPressed: () async {
-          await showDialog<bool>(
-            context: context,
-            builder: (context) => Center(
-              child: SingleChildScrollView(
-                child: MyAlertDialog(
-                  title: Text(Language.of(context).message),
-                  content:
-                      Text(Language.of(context).deleteAllNotesResetPassword),
-                  actions: [
-                    TextButton(
-                      onPressed: () =>
-                          resetPassword(context, deleteAllNotes: true),
-                      child: Text(Language.of(context).alertDialogOp1),
+          final status = await showDialog<bool>(
+                barrierDismissible: false,
+                context: context,
+                builder: (context) => Center(
+                  child: SingleChildScrollView(
+                    child: MyAlertDialog(
+                      title: Text(Language.of(context).message),
+                      content: Text(
+                          Language.of(context).deleteAllNotesResetPassword),
+                      actions: [
+                        TextButton(
+                          onPressed: () {
+                            Navigator.of(context).pop(true);
+                          },
+                          child: Text(Language.of(context).alertDialogOp1),
+                        ),
+                        TextButton(
+                          onPressed: () async {
+                            Navigator.of(context).pop(false);
+                          },
+                          child: Text(Language.of(context).alertDialogOp2),
+                        ),
+                      ],
                     ),
-                    TextButton(
-                      onPressed: () async {
-                        Navigator.of(context).pop(true);
-                      },
-                      child: Text(Language.of(context).alertDialogOp2),
-                    ),
-                  ],
+                  ),
                 ),
-              ),
-            ),
-          );
+              ) ??
+              false;
+          if (status) {
+            await resetPassword(context, deleteAllNotes: true);
+          }
         },
       );
 
@@ -178,15 +183,15 @@ class Utilities {
       );
 
   static Future<void> onHideTap(BuildContext context, Note note) async {
-    final status = Provider.of<LockChecker>(context, listen: false).passwordSet;
+    final status =
+        Provider.of<LockChecker>(context, listen: false).password.isNotEmpty;
     if (!status) {
       await showDialog(
+        barrierDismissible: true,
         context: context,
-        builder: (_) => MySimpleDialog(
-          title: Padding(
-            padding: const EdgeInsets.all(8),
-            child: Text(Language.of(context).setPasswordFirst),
-          ),
+        builder: (_) => MyAlertDialog(
+          title: Text(Language.of(context).message),
+          content: Text(Language.of(context).setPasswordFirst),
         ),
       );
     } else {
@@ -221,6 +226,7 @@ class Utilities {
     var choice = true;
     if (!deleteDirectly) {
       choice = await showDialog<bool>(
+        barrierDismissible: false,
             context: context,
             builder: (_) => MyAlertDialog(
               title: Text(Language.of(context).message),
